pyproject.toml
[tool.poetry]
name = "chilla-myid"
version = "1.0"
description = "Chilla is the daughter of Doddi Priyambodo"
authors = ["Doddi Priyambodo <doddi@bicarait.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.33.0"
google-cloud-aiplatform = "^1.48.0"
google-cloud-logging = "^3.10.0"
langchain = "^0.1.16"
pytz = "^2024.1"
google-cloud-discoveryengine = "^0.11.11"
transformers = "^4.40.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
